LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/all-with-discrete.tentative-expected.txt

hello

FAIL transition:all with transition-behavior:allow-discrete should animate discrete properties. assert_equals: Did not start a discrete-property transition expected 1 but got 0
